+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".cart.CartActivity"
+            android:name=".view.cart.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".productdetail.ProductDetailActivity"
+            android:name=".view.productdetail.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".productcatalogue.ProductCatalogueActivity"
+            android:name=".view.productcatalogue.ProductCatalogueActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
