+++ app/src/main/java/woowacourse/shopping/data/BundleKeys.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.data
 
 object BundleKeys {
     const val KEY_PRODUCT = "product"
