+++ app/src/main/java/woowacourse/shopping/data/ProductMockData.kt
@@ -1,9 +1,10 @@
-package woowacourse.shopping
+package woowacourse.shopping.data
 
 import com.shopping.domain.MainProductCatalogue
 import com.shopping.domain.Product
 import com.shopping.domain.RecentProductCatalogue
-import woowacourse.shopping.mapper.toUIModel
+import woowacourse.shopping.toUIModel
+import woowacourse.shopping.uimodel.mapper.toUIModel
 
 object ProductMockData {
     val products = listOf(
