+++ app/src/main/java/woowacourse/shopping/data/db/CartDBHelper.kt
@@ -1,10 +1,10 @@
-package woowacourse.shopping
+package woowacourse.shopping.data.db
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 
-class ProductDBHelper(context: Context) : SQLiteOpenHelper(
+class CartDBHelper(context: Context) : SQLiteOpenHelper(
     context,
     DB_NAME,
     null,
@@ -19,7 +19,7 @@ class ProductDBHelper(context: Context) : SQLiteOpenHelper(
     }
 
     companion object {
-        const val DB_NAME = "ProductDB"
+        const val DB_NAME = "CartDB"
         private const val DB_VERSION: Int = 1
         const val TABLE_NAME = "products"
         const val KEY_NAME = "name"
