+++ app/src/main/java/woowacourse/shopping/uimodel/MainProductCatalogueUIModel.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping
+package woowacourse.shopping.uimodel
 
 class MainProductCatalogueUIModel(val items: List<ProductUIModel>)
