+++ app/src/main/java/woowacourse/shopping/uimodel/mapper/MainProductCatalogueMapper.kt
@@ -1,8 +1,9 @@
 package woowacourse.shopping
 
 import com.shopping.domain.MainProductCatalogue
-import woowacourse.shopping.mapper.toDomain
-import woowacourse.shopping.mapper.toUIModel
+import woowacourse.shopping.uimodel.MainProductCatalogueUIModel
+import woowacourse.shopping.uimodel.mapper.toDomain
+import woowacourse.shopping.uimodel.mapper.toUIModel
 
 fun MainProductCatalogueUIModel.toDomain(): MainProductCatalogue =
     MainProductCatalogue(items.map { it.toDomain() })
