+++ app/src/main/java/woowacourse/shopping/uimodel/mapper/ProductMapper.kt
@@ -1,7 +1,7 @@
-package woowacourse.shopping.mapper
+package woowacourse.shopping.uimodel.mapper
 
 import com.shopping.domain.Product
-import woowacourse.shopping.ProductUIModel
+import woowacourse.shopping.uimodel.ProductUIModel
 
 fun ProductUIModel.toDomain() =
     Product(url, name, price)
