+++ app/src/main/java/woowacourse/shopping/uimodel/mapper/RecentProductCatalogueMapper.kt
@@ -1,10 +1,10 @@
-package woowacourse.shopping.mapper
+package woowacourse.shopping.uimodel.mapper
 
 import com.shopping.domain.MainProductCatalogue
 import com.shopping.domain.RecentProductCatalogue
-import woowacourse.shopping.RecentProductCatalogueUIModel
 import woowacourse.shopping.toDomain
 import woowacourse.shopping.toUIModel
+import woowacourse.shopping.uimodel.RecentProductCatalogueUIModel
 
 fun RecentProductCatalogueUIModel.toDomain() =
     RecentProductCatalogue(mainProductCatalogue.toDomain())
