+++ app/src/main/java/woowacourse/shopping/view/cart/CartRecyclerViewHolder.kt
@@ -1,8 +1,10 @@
-package woowacourse.shopping
+package woowacourse.shopping.view.cart
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import woowacourse.shopping.databinding.ItemProductInCartBinding
+import woowacourse.shopping.uimodel.CartUIModel
+import woowacourse.shopping.uimodel.ProductUIModel
 
 class CartRecyclerViewHolder(
     private val binding: ItemProductInCartBinding,
