+++ app/src/main/java/woowacourse/shopping/view/productdetail/ProductDetailActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.productdetail
+package woowacourse.shopping.view.productdetail
 
 import android.content.Context
 import android.content.Intent
@@ -8,15 +8,15 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.bumptech.glide.Glide
-import woowacourse.shopping.BundleKeys
-import woowacourse.shopping.CartDBHelper
-import woowacourse.shopping.ProductDBHelper
-import woowacourse.shopping.ProductDBRepository
-import woowacourse.shopping.ProductUIModel
 import woowacourse.shopping.R
-import woowacourse.shopping.cart.CartActivity
+import woowacourse.shopping.data.BundleKeys
+import woowacourse.shopping.data.db.CartDBHelper
+import woowacourse.shopping.data.db.ProductDBHelper
+import woowacourse.shopping.data.db.ProductDBRepository
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.getSerializableCompat
+import woowacourse.shopping.uimodel.ProductUIModel
+import woowacourse.shopping.view.cart.CartActivity
 
 class ProductDetailActivity : AppCompatActivity() {
     private lateinit var binding: ActivityProductDetailBinding
