+++ app/src/main/res/layout/activity_product_catalogue.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".productcatalogue.ProductCatalogueActivity">
+        tools:context=".view.productcatalogue.ProductCatalogueActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tb_product_catalogue"
