+++ app/src/main/res/layout/activity_product_detail.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".productdetail.ProductDetailActivity">
+        tools:context=".view.productdetail.ProductDetailActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tb_product_detail"
